[reconfig] Add more logging (#7929)
This PR adds more logging related to reconfig.
One major change is to be able to log the longest task while waiting for
effects to be available. This will provide useful information when we
want to debug the delay between when we want to create checkpoint and
when it's finally created.