[2/x][adapter] Transactional tests include genesis objects
Add genesis objects to the transactional test runner so that we can
test functions involving them in isolation.  Start by adding the
`Clock`.

This change bumps the Fake IDs of every object in existing tests by
one, so the bulk of this change is updating those.

Test Plan:

```
sui-adapter-transactional-tests$ cargo nextest run
```