Write certificates to pending_certificates table in SuiTxValidator (#8300)
Currently `TransactionManager` loads certificates from
`pending_certificates` table when they are read to be executed. So it is
always necessary to writes certificates to the table before enqueueing
them to `TransactionManager`. We can consider keeping certificates in
memory in `TransactionManager`, and only persist certificates from
Narwhal output. But that would be another change.