[epoch_data] Define EpochData and pass it to transaction context (1/n)
This PR adds new data structure, `EpochData` and passes it to the transaction context.

This data structure is the bridge between sui's internal `EpochStartConfiguration` and the `TxContext` that is passed to every move call and allows to specify epoch-level data all move calls at a zero cost.

Future PR will add more fields to `EpochData`.