Ability to change current page via progress bar When a user creates new transaction and wants to go to the previous screen to change amount, he must use back button which is small and is not convenient. Also the user can think that pressing on the back button will drop him to the balance page. I think it would be better to allow pressing on a progress bar at the top.
So, for example, when you entered a transaction amount and pressed `Continue`, you may press on the `Amount` tab at the top which will show the first page with amount selection.

Pressing on the next pages in the progress bar will do nothing since data on pages should be verified first.