[hardening] towards checking SUI conservation across each transaction
The SUI supply is fixed at 10B as of genesis, so transactions should never create/destroy SUI.
This PR takes a step toward writing dynamic checks for helping to ensure this by introducing a function that computes the total amount of SUI stored in an object.
This function is used by `temporary_store` to compute the amount of SUI in input objects, and then ensure that the amount of SUI in output objects + the gas fees match.

The intention is to call this inside `adapter::execute_transaction` after gas charging has occurred, but there are some Move trait problems that prevent me from easily doing this at the moment--opened https://github.com/move-language/move/pull/935 to help with this