[upgrade] Add ability to upgrade system state at epoch change (#8763)
This PR adds the following things:
1. Added a function that maps from protocol version to system state
version. This allows us to decide what would be the new system state
version in the new epoch.
2. Pass the system state version in the advance epoch transaction, and
if it changes, we update the version and also upgrade the system state
object. The actual upgrade doesn't happen today, and that will only be
added when we have a real update. We will need to figure out a way to
upgrade it for testing.