[gas] fix SUI conservation issue by refactoring gas charging logic
- Encapsulate all computation gas, storage gas, rebate logic, and "reset store due to out of gas + redo gas-smashing logic" inside `temporary_store::charge_gas`. This makes it easier to debug and change things.
- Separate charging gas for storage and gas for computation, and ensure that charging gas for storage never fails. This is important because if we run out of gas while charging for storage, we want to roll back all effects of *attempting* to charge for storage. This is infeasible if storage/computation cost for object mutations is entangled, as it currently is. Instead, we first charge gas for storage and track the total size of the storage changes, then charge for computation gas at the end.

If we don't use this discipline, there are several scenarios in which we can accidentally create or destroy SUI, e.g.,
- Add X to the tx storage rebate, fail while charging storage gas for X. Effects on the storage rebate aren't reverted, X SUI gets created

I found this via running https://github.com/MystenLabs/sui/pull/8858 on the existing tests. The test that exhibits this behavior is `sui-transactional-tests/sui/freeze.move`, but it would be great to add a dedicated unit test for this case if someone familiar with the gas tests can point me to where we test the effect of changing the storage gas price.