Delete fetchAllOwnedAndRequiredObjects and related file (#8853)
## Description 

Describe the changes or additions included in this PR.

This PR closes the Wallet coin refactor. Deleting
`fetchAllOwnedAndRequiredObjects` and any redux dependency

- Remove `fetchAllOwnedAndRequiredObjects` and related dependency 
- Move mint NFT inline 
- Deleted related/unused  files 

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed and
then add it to the Release Notes section as highlighted during the
release process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------

Co-authored-by: Jibz <godwinjibs@gmail.com>