change the coin fetching api in active-coin-card (#8824)
Describe the changes or additions included in this PR.

- Added useGetCoins - gets all coins for a given address
- `selectCoinSetWithCombinedBalanceGreaterThanOrEqual`, `sortByBalance`
and `totalBalance` uses `CoinStruct` instead of `SuiMoveObject`
-  `getBalanceFromCoinStruct`
- Updated ActiveCoinsCard to use `useGetAllBalances`

How did you test the new or updated feature?

`ActiveCoinsCard` uses `useGetAllBalances`

https://user-images.githubusercontent.com/126525197/222857161-47f4c4af-34b0-4209-8185-f028e98d8a03.mov

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

---------

Co-authored-by: Jibz <godwinjibs@gmail.com>