[programmable transactions] Add support for publishing `UpgradeCap` during transactional tests (#8981)
This adds the `--upgradeable` flag to transaction tests that, when
present, will transfer the returned `UpgradeCap` that is returned from
publishing a package to the sender of the transaction.

This adds a new entrypoint
`ProgrammableTransaction::publish_and_handle_capability_with` to support
this -- you can pass a capability handler which given a capability will
return a command (and may perhaps add more commands) that will be placed
immediately after the publish command.

This PR is stacked on #8825

Added a new transactional test with the `--upgradeable` flag that makes
sure the `UpgradeCap` is published under the sender of the transaction.