[upgrades] Fix move_integration_test "Not a package" panic
This panic happened because `test_publishing_with_unpublished_deps`
assumed that the first object created by package publishing would
always be the package, but in fact sometimes it is the `UpgradeCap`.

Fixed by augmenting an existing helper function that can correctly
distinguish the two created objects so it can be used in the
"unpublished deps" case.

Test Plan:

```
$ cargo nextest run -- 'test_publishing_with_unpublished_deps'
```