[Sui] stop waiting for transaction submission once it is processed (#9189)
## Description 

We have observed transactions stuck in consensus adapter likely because
the submission to Narwhal is stuck in one of the channels or buffers
before it is acknowledged by Proposer. It seems better to cancel the
submission once the transaction is out of consensus. This would allow
new transactions to be submitted to Narwhal, and Narwhal can implement
load shedding internally to drop old transactions.

If this can be done at a higher level, it would be a good alternative
too.

## Test Plan 

Deployed to private testnet.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes