Use checkpoint execution to detect stuck transactions in TransactionManager (#9250)
## Description 

Since transactions are causally ordered in checkpoints, executing
transactions sequentially in checkpoint should encounter no missing
input object. We can rely on this to detect and print diagnostic
information for stuck transactions in TransactionManager: if the stuck
transaction is already part of a checkpoint, its missing input objects
should be printed to help debug why the transaction is not ready for
execution.

However this change does not help when a transaction cannot execute at
any node and is not part of a checkpoint. We can add debug logging in
checkpoint builder if needed.

## Test Plan 

Tried it out in a test with stuck transactions.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes