Set an upper limit to the number of public inputs for the Groth16 verifier (#9336)
## Description 

To avoid overloading validators, we need to set an upper limit to the
number of public inputs for the Groth16 verifier.

## Test Plan 

Pass unit tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [X] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Set an upper limit to the number of public inputs to the Groth16
verifier in the Move API. The upper limit is currently set to eight
inputs.