Add ValidatorWrapper and use it in inactive_validators (#9328)
This PR introduces a new Move type called Versioned. It represents a
wrapper of a versioned inner type. This provides flexibility in what
type to load from the inner (dynamic field) object.
We then define ValidatorWrapper which is a thin wrapper of Versioned to
provide all interfaces that hide the inner type of Validator.
This is then used in `inactive_validators`. We need this because it's
impossible to migrate all `inactive_validators` to use new types when we
upgrade the Validator type, and hence we must support this form of
polymorphism.
In a follow up PR I will also use this type for validator_candidates.