[Narwhal] swap typed_store::Store with DBMap (#9387)
## Description 

This PR swaps for Narwhal all the `typed_store::Store` instances with
`DBMap`. Also, I've wrapped the `DBMaps` behind dedicated structs so we
can make things easier if we consider to swap later on `DBMaps` with
something different in the future + make it easier to add additional
methods on each store if needed in the future (ex update additional
secondary index stores in batch fashion etc).

## Test Plan 

Existing & new unit and integration tests.

## TODO:
- [ ] wrap the `batch_store` DBMap behind a dedicated struct

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes