Fixes for gas (#9449)
## Description 

Correct a couple of problems with gas. One of the change is gated by
feature flag and requires a protocol version.
So let's make sure to check that code carefully and see the tests that
had to change too.
The other change is going in without protocol version and it's explained
in a comment. I would love to get feedback on that.

Let me know what else you would like to see here

## Test Plan 

Existing tests. 
The indexer failure is unrelated to this change and fixes are being
deployed

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes