[hardening] dynamic checks for ensuring all objects read by a tx have been authenticated
One of the key security properties of a Sui transaction is that it should only read objects that have been authenticated: e.g., the object is owned directly by the sender, is shared or immutable, or is indirectly owned by the sender or an input object.

This adds a dynamic check at the end of tx execution to ensure that every object accessed can be authenticated w.r.t one of the roots. For now, we gate this to run in debug mode only.