Make change in signing a versioned operation (#9517)
## Description 

Protect the fix for signing with a version check.
The worry here is that while the system is upgrading a version
unprotected behavior may lead to different results in signing. If that
should not cause forking it may create equivocation and objects locked
for an epoch

## Test Plan 

Existing tests and writing few specific ones

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes