End friendships in framework code using programmable txns (#9509)
## Description 

This PR removes the friendship between modules in `governance` folder
and modules that are not. This is preparation for separating
`governance` into its own package.

A few things are changed to make this happen:
- Some `public(friend)` functions are made private
- The signatures of genesis and epoch change function are changed
- Genesis and epoch change are now multi-step programmable txns executed
in `System` mode


## Test Plan 

Making sure existing tests pass.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [x] breaking change for FNs (FN binary must upgrade)
- [x] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes