[sui-framework] Split transfer functions into public and internal variants (#9498)
## Description 

- transfer, share_object, and freeze_object are now internal to the
module.
- Added public_ variants that are not but require store.
- This split makes the existing behavior more transparent, and allows us
to update the code in the future if/when we get new type system
features.

## Test Plan 

- Migrated tests
- New tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [X] breaking change for a client SDKs
- [X] breaking change for FNs (FN binary must upgrade)
- [X] breaking change for validators or node operators (must upgrade
binaries)
- [X] breaking change for on-chain data layout
- [X] necessitate either a data wipe or data migration

### Release notes

- Before this change, `transfer::transfer` (and other `transfer`
functions) were effectively overloaded. They could be called either to
interact with an object in its module OR with any object that has the
`store` type ability.
- We received a lot of feedback that this was confusing. And
furthermore, if we wanted to generalize this to possibly adding an
`internal` constraint on type parameters to make this check more
generalized.
- After this change, `transfer::transfer` has been split into two
functions: `public_transfer` which requires `store`, and `transfer` that
does not but can only be used in the object's module. Similar splits
have happened for `share_object` and `freeze_object`.