[ConsensusAdapter] make max pending transactions configurable (#9640)
## Description 

This will allow different parameters to be configured on different
environments.

Also the limit is raised from 40k to 100k, but I can also keep the
original limit and only raise it for private testnet.

## Test Plan 

Existing unit tests. Later deploy to an environment.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes