[package upgrades] Add fast-path package upgrade CLI command (#9594)
This adds a fast-path upgrade for the Sui CLI I specifically didn't want
to add it to the transaction builder/json-rpc API at the moment since we
will probably want to to think that API through quite a bit more so we
don't have to make too many breaking changes to it (and to allow custom
upgrade policies).

Added a new test (there's some nastiness around `published-at` that's
needed in there sadly), and exercised the upgrade path locally in the
CLI as well.