[wallet-ext] refactor the importing of Ledger accounts to use useQuery (#9661)
## Description 
This PR refactors the importing of Ledger accounts to use `useQuery`
which is a little more ergonomic than using `useEffect`. In the rush of
meeting last week's deadline, I overlooked that we can use `useQuery`
for any type of async data retrieval operation (the "server" in this
case is the Ledger device). This reduces some of the boilerplate code
for loading/error states and when caching is possible at some point,
we'll get it out of the box for free.

## Test Plan 
- Manual testing (initial connection, cached re-connection,
disconnection during the middle of the flow, etc.)

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A