[wallet-ext] re-factor LedgerSigner to handle lazily connecting to Ledger and do some other clean-up type work (#9635)
## Description 
This PR refactors the Ledger signing flow so that `useSigner` doesn't
return an `async` function. As I was prepping this for multi-Ledger
support, I also figured that it would be better to always prompt the
user to select what Ledger device they'd like to sign transactions with
instead of just connecting to the first connected device.

## Test Plan 
- Manual testing (signed/staked/approved transaction requests with and
without a Ledger account)

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A