add more functions for valdiator tool (#9626)
This PR adds more functionalities in the validator cli, including:
1. set gas
2. update metadata
3. report validators
4. display metadata

**This is only available for pending/active validators today, because
there's a breaking change in candidate/inactive, so let's wait.**

It also updates the default genesis gas amount to 30M SUI (for local
network), so it's much easier to test staking flow.

Test all the commands locally.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

More stuff to validator tool