package publish: error for unpublished deps with nonzero addrs (#9254)
## Description 

Error if we see that package dependencies have self-addresses other than
`0x0` on `publish` and advise action.

## Test Plan 

Added test.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

If a user attempts to `publish` a package where the package's
dependencies are set non-zero addresses _and_ the user sets
`--with-unpublished-dependencies` in the command line, we error and
advise action: either set dependency addresses to `0x0` (the zero
address) or remove `--with-unpublished-dependencies` after ensuruing
that dependencies are published.