Ts-sdk: change number to bigint or string in provider functions (#9412)
## Description 

- Function `getTotalTransactionNumber` of ts-sdk now returns a `bigint`.
- Function `getReferenceGasPrice` of ts-sdk now returns a `bigint`.
- `gasPrice` of `devInspectTransaction` is defined as a `bigint`.
- `getLatestCheckpointSequenceNumber` returns a `string`
- `getCheckpoint` takes as input a `string` for checkpoint sequence
number.
- `getCheckpoints` takes as input a `string` for cursor.
- Introduce `SuiCheckpointSequenceNumber` type to sui-json-rpc-types
that is defined as `BigInt` (compared to `CheckpointSequenceNumber` used
from sui-types that was defined as a `u64`)
- Change rust side to support the above. Changes to explorer, sui-sdk,
sui-indexer and sui-rosetta that got affected by the changes.

PR #9347 was adding a lot of changes, so I am splitting it into multiple
PRs so that the changes can be better reviewed and tested. A related PR:
#9462

## Test Plan 

manually test explorer and wallet

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------