[refactor] introduce an LRU cache inside certificate store (#9760)
## Description 

This PR is introducing an LRU cache in the certificate_store in order to
minimise the impact from the read operations (only on the primary index
at the moment). That will help:
* reduce the number of reads against rocksdb
* avoid deserialisation costs of Certificate 

relevant metrics for cache hit/miss have been added.

The cache has been set to hold 100 * 60 items (100 nodes X 60 rounds) -
but we can tune further. Also, we only populate the cache when we write
a certificate - especially for the cases we are interested I don't
consider we need to re-populate on read.

## Test Plan 

Added unit tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes