Gas cleanup again (#9806)
## Description 

This is pretty simple and mostly a cleanup.
It also adds a gas model version to the protocol config so we can easily
fork the code.
There are pretty much only 2 entry points to protect (`SuiGasStatus`
creation and `charge_gas` call) to version gas properly.
And we know we need to version super soon.
The work I was supposed to land is getting too intrusive and adding the
handling of system transaction is way too much for today.
I don't feel comfortable with any work in that space going in today, I'd
rather fix things better over the weekend and push the change in an
upgrade as soon as we can next week.
@lxfind @bmwill @emmazzz @sblackshear

## Test Plan 

Existing tests at the moment, as this is not a new feature but a core
change. More tests would be nice soon

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes