[move-package] Create MovePackage before verification
Preparation for package upgrades, where to verify `CompiledModules`,
we need linkage and type origin tables, which are created as part of a
`MovePackage`, but currently, the `MovePackage` is created after
verification.

This change allows us to deserialize bytes to create
`CompiledModule`s, then build a `MovePackage` from those by reference,
and use that to publish and verify the deserialized modules.

Removes version parameter from MovePackage::new_initial

Get rid of `publish_and_verify_new_modules`,
`publish_and_verify_upgraded_modules` and `generate_package_id` in
favour of inlining them into their respective call-sites.  This is to
prepare for generating a `MovePackage` before calling
`publish_and_verify_modules` which requires having access to the list
of `CompiledModule`s after deserialization and address substitution.

In preparation for verification requiring a context that needs to be
derived from a `MovePackage`.

Test Plan:

```
$ cargo simtest
$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```