[linkage] Set linkage table for ExecutionContext
WIP: Add an API to set the linkage table based on a `MovePackage`, and
read from it when responding to `relocate` and `link_context` queries.

The `LinkageView` now expects a link context to be set -- in turn,
calls into the session that load modules will fail because we aren't
setting the link context correctly yet.

Test Plan:

Existing tests + newly enabled upgrade tests:

```
$ cargo simtest
$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```