WIP: Set linkage context for move calls
Initial work on setting the appropriate linkage context for calls into
the VM.  Support for setting the correct context for loading
types (including type parameters for calls) is TBD, so not all tests
will pass.

To help identify issues, this PR also makes it no longer an error to
run code without a linkage context set -- this provides basic support
for loading types in the absence of upgrades, but will need to be
changed back.

Test Plan:

(9 tests failing at the moment, mostly because of issues related to
type parameters).

```
$ cargo simtest
$ cargo nextest run
```