[upgrades] Set link table
Initial work on setting the appropriate linkage context for calls into
the VM and type loading.

To help identify issues, this PR also makes it no longer an error to
run code without a linkage context set -- this provides basic support
for loading types in the absence of upgrades, but may need to be
changed back.

Test Plan:

```
$ cargo simtest
$ cargo nextest run
```