[transaction-builder] Fix place where we were treating gas budget as units instead of Mist
This fixes an issue where we were treating the gas budget in the transaction builder as representing units of gas instead of Mist.

This fixes this issue and also adds an error if the provided gas budget is below the current RGP (and therefore equivalent to 0 gas units after the div-truncation that is done when entering execution and converting to gas units).