[future proof/hardening] Hide internals of MoveObjectType (#9964)
## Description 

Made the enum for MoveObjectType private to avoid manually constructing
an Other. This is potentially an issue as the type does not support
equality currently over similar types, e.g. `GasCoin !=
Other(GasCoin::type_())`.
In other words, this is more about trying to make sure buggy code isn't
introduced in the future.

## Test Plan 

- Internal change 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes