[programmable transactions] LinkageView
Remove implementation of LinkageResolver for `StorageView`, and
introduce `LinkageView` as a wrapping type that holds linkage
information and delegates all other resolution queries to
`StorageView`.

Relies on move-language/move#1009 which adds support to `Session` for
resolvers that are richer than just a reference.

Currently, the implementation of `LinkageResolver` does no re-linking,
this PR only introduces the abstraction.

Test Plan:

```
$ cargo simtest
$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```