[StateAccumulator] Bug fix (#10108)
## Description 

A potential bug exists due to how we process multiple transaction
effects from the same checkpoint. Our treatment of
`unwrapped`/`unwrapped_then_deleted` effects is, rather than find the
object metadata in `modified_at_effects`, to lookup the `ObjectRef` info
from the object store directly for the correct object version (as the
one in `modified_at` will not give us the version that we added to the
accumulator during the previous wrap (as a tombstone).

However, to determine which ids should be skipped and processed later as
shown above, we create a HashSet of all unwrapped objects from all txes
in the checkpoint. As such, we potentially skip over removals that
should not be that may have happened in a different checkpoint (as if
there is a future unwrap, then the id will be in our set, and any other
txes where that id shows up in `modified_at_versions` will be ignored).

The fix here is to make this ignore set tx specific, and only check the
unwrapped objects for the corresponding tx when looking through
`modified_at_versions`.

## Test Plan 

```
cargo simtest && cargo nextest run move_integration_tests
```

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes