Update fastcrypto - includes an update to ECVRF which is not backward compatible with previous ECVRF proofs. (#10332)
## Description 
Update fastcrypto - includes an update to ECVRF which is not backward
compatible with previous ECVRF proofs.

## Test Plan 
All tests pass

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [x] breaking change for a client SDKs
- [x] breaking change for FNs (FN binary must upgrade)
- [x] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Update fastcrypto - includes an update to ECVRF which is not backward
compatible with previous ECVRF proofs.