[pruner] fix last executed checkpoint watermark bug (#10372)
In an aggressive setup `(num_epochs_to_retain=0)`, the pruner runs
constantly and deletes modified/deleted versions of objects that were
part of some executed checkpoint.
It relies on the `HighestExecuted` watermark to avoid pruning too much. 
The `checkpoint_number` variable needs to be set before the check;
otherwise, the pruner might process the next checkpoint.