[StateAccumulator] Check consistency in expensive mode (#10290)
## Description 

Checking for state consistency is too expensive to be run blocking
between epochs in general. So either we run in the background or we make
this check opt-in only.

Option (1) may be introduced later via RocksDB snapshots, but for now
this PR introduces option (2).

## Test Plan 

Existing tests since this will run in expensive mode or debug mode.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes