[publish] Preserve execution order for publish in backward compat (#10414)
## Description 

Fixing publish transaction backwards compatibility (introduced at the
time of upgrades). The backwards compatible implementation changed the
order of operations (execute `init` functions before fetching packages),
which resulted in a change in the amount of gas charged, because we were
executing instructions on one path and not on the other.

## Test Plan 

```
$ cargo simtest
$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```

And run a fullnode on devnet and ensure it catches up past checkpoint
32314 (the checkpoint that was previously failing).
---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [x] breaking change for FNs (FN binary must upgrade)
- [x] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes