[JSONRPC] - offload jsonrpc call to todo blocking thread for method with large response (#10438)
jsonrpsee server occasionally refuse to process any request when the
request rates are high, after some investigation we narrowed down the
problem to json serialisation. When the response object is large, the
serde takes significant amount of time and blocks the tokio worker
threads, currently the only way to offload the serialisation is to make
the RPC method "blocking" and jsonrpsee will use tokio spawn_blocking to
process the request.

This PR marked all methods with relatively large response "blocking" to
avoid blocking

RPC endpoints are covered by integration tests