[bug] check that all gas coins have object owners
The tx validation logic checks previously checked that the first gas coin had an address owner, but not the others. Passing shared or object-owned gas coins worked correctly, but passing an immutable gas coin
leads to this object getting deleted by gas smashing, which is obviously not intended.

This PR fixes the issue and adds some hardening for deletion of immutable objects via runtime checks.

Credit to Alexis at Halborn for this finding.