[consensus] backwards compatible consensus changes (#10508)
## Description 

This PR is making the changes of
https://github.com/MystenLabs/sui/pull/10507 backwards compatible. More
specifically it is:
* introducing a new table in consensus store that will store from now on
all the Committed Subdags in a versioned format. I've decided to follow
the same approach we have used in other parts of narwhal (ex Header,
Certificate etc) for consistency reasons, so we don't mix and match
approaches.
* making the current timestamp changes backwards compatible by
introducing V2
* small refactorings (ex moved the ConsensusStore under the `storage`
module where it should be)

## Test Plan 

Existing & new unit/integration tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes