[faucet] Avoid cloning WAL DBMap when retrying
The clone was to get around the fact that we needed to increment a
count in the log while iterating over it, but `DBMap`s contain metrics
etc, so it's probably not a good idea to clone them.  Get around this
by going over the pending coins in a separate pass to increment the
retry count.

Test Plan:

```
crates/sui-faucet$ cargo nextest run
```