Remove dummy gas price and robustify gas budget in tests (#10510)
For Reviewers: meat is in non-test files.

Today the gas budget in our tests are hardcoded and does not adapt to
dynamic gas prices. This has caused several problems in the past. This
PR is the first step to robustify the gas budget in our tests. In
particular, this PR:
1. remove `DUMMY_GAS_PRICE` and all functions using it
2. use reference gas price and GAS_UNIT ffwhen making transactions
3. change the default reference gas price to `100`
4. fix a few flaky tests

Caveat:
1. This change _may_ break load tests
2. So far the `GAS_UNIT` is conflated with storage price, which is
costly for genesis objects. In a higher gas price setting e.g. `1000`,
some tests would fail because the Gas is too high.

Tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration