[Compatibility] Fix debug log level selection logic; Increase ssh ttl (#10789)
## Description 

... also, introduce `clean` option, which defaults to true. When
`false`, will not clean the database. This is helpful for kicking off
another run from the previous state of a bad run, rather than running
from clean state and having to wait many hours to reach the previous
failpoint.

## Test Plan 

Ran with `pull_request` scheduling added in order to check in CI (and
later removed). Verified that in the CI run, we were not running with
`DEBUG` log level

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes