[checkpoints] Detect checkpoint hard fork early
This PR detects checkpoints hard fork early in checkpoint builder and panics validator. This is needed to prevent chain from signing certificates on the fast path, reducing potential fork impact.

There is some discussion whether panic is the right response in this case. This PR follows pattern established in the `CheckpointExecutor` and panics on fork. We can revisit both places in a separate PR if we want different response in case of a fork.