[move build] example linter: transfer-to-self
Warn on calls of `public_transfer(o, tx_context::sender())`, and advise the programmer to return `o` from the function instead. This nudges folks toward writing compositional code that can be used in programmable transaction blocks.

We do not warn on `transfer(o, tx_context::sender())`, or in `public_transfer(o, addr)` where addr might not be `tx_context_sender()`. In such cases, the programmer might not be able to follow the advice of returning the object.