Avoid writing owned object transactions into pending_execution table from RPC handler (#11014)
## Description 

Transactions submitted via `handle_certificate()` are persisted into the
`pending_consensus_transactions` table in consensus adapter, and this
ensures the transactions will be propagated to other nodes. It is
unnecessary to persist the transaction into the pending_execution table
again inside the RPC handler.

## Test Plan 

Existing crash recovery tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes