[sui-framework] Remove build_move_package
This function was responsible for building a package and then
verifying the version of the framework they were built against, but
the verification had been commented out and since superseded by source
validation in general.

This change allows us to remove a dependency on `sui-framework` from a
couple of crates that were only using it for builds.

Test Plan:

```
$ cargo simtest
$ env SUI_SKIP_SIMTSTS=1 cargo nextest run
```