[sui-framework-build] Use ObjectIDs to distinguish system packages
...rather than their package names, this allows getting rid of package
names declared in `sui-framework-build/lib.rs`, the last vestige of
framework-specific code from `sui-framework-build`.

Also rename `SUI_SYSTEM_PACKAGE_ID` to `SUI_SYSTEM_OBJECT_ID` for
consistency with the rest of the system package ID constants.

In a follow-up PR, we will then rename `sui-framework-build` to
`sui-move-build` and consolidate it down to one module, because that
is what it is for (not just for building the framework).

Test Plan:

```
$ cargo simtest
$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```