[sui-types] Move convert_vm_error to sui-adapter
`convert_vm_error` is only used within `sui-adapter`, and introduces a
dependency on `move-vm-runtime` (which will be versioned with the
execution layer) to `sui-types` (which will not).

Moving `convert_vm_error` to `sui-adapter` removes the dependency and
brings the function closer to its call-sites which also improves
encapsulation.

Test Plan:

```
$ cargo simtest
$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```