[rfc][adapter optimization] eliminate redundant serialization and deserialization of coin values
`Object::new()` shows up in the flamegraph, and it does some redundant serialization of coin values into the `Coin` type. This PR eliminates usage of the `Coin` type inside the adapter and instead operates directly on the raw bytes, eliminating any serialization/deserialization overhead. We've seen this sort of optimization make a difference elsewhere (e.g., in the conservation checks), so it's worth trying here.

This should be ready to test for perf purposes, but it should be tested/reviewed carefully both for correctness and backward compatibility before we consider landing--it is touching very sensitive logic.