[sui verifier] Ban entry on init (#11287)
## Description 

- No longer allow entry modifier on init
- `entry` on `init` while not harmful, is a bit confusing as `init`
functions are otherwise not callable. For cleanliness, we think it is
important to keep it separated as one might expect `init` is only called
once
- Thanks @otter-sec (https://osec.io/) for the report! 

## Test Plan 

- New tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes