Placing docs in subdirectory to prevent overwriting with every publish (#10782)
## Description 

Places crate documentation in a sub-directory and removes the
force_orphan option so the branch does not clear with every publish. A
redirect will be placed in the gh-pages branch to direct users
appropriately.

## Test Plan 

Local

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes