[config] Introduce LiveConfig
This introduces LiveConfig that automatically refreshes config when receiving a SIGHUP signal.

The intended usage is that caller holds a reference to `LiveConfig<T>` and calls `LiveConfig::latest()` when it needs to read config value. The value in `latest()` is atomically swapped when new config is available.

We can introduce more APIs(such as subscribing to change) later if we have use cases for it.

Note on the signal:
* Just updating config file is not enough, user still needs to manually send SIGHUP signal via something like `killall sui-node -SIGHUP`
* SIGHUP is conventionally commonly used by daemons as a signal to reload configuration: https://stackoverflow.com/questions/19052354/sighup-for-reloading-configuration
* Since signals are only available on unix, on the non-unix system we simply refresh the file every 30 seconds
* We can consider external file watcher libraries like `notify-rs`, but their usage does come with caveats, and it seems that explicit SIGHUP signal might be less error-prone.