[ConsensusAdapter] reduce the min expected consensus latency (#11547)
## Description 

p50 consensus latency is < 2s on testnet, so min resubmission latency
should be lower, e.g. 2s * 1.5 ~ 3s, instead of 5s * 1.5 ~ 7.5s. If the
p50 latency increases for various reasons, we rely on `LatencyObserver`
to provide a more accurate value.

## Test Plan 

CI. Private testnet.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes