Take db snapshot after epoch start config is added to db (#11700)
## Description 

Before this PR, we would snapshot before epoch start config is added for
the next epoch. Which would mean a restored snapshot would start in the
current epoch and then reconfigure (and run accumulator again which
breaks with aggressive snapshot because we also aggressively prune the
snapshots).

## Test Plan 

Updated integration test to use aggressive pruning.