[Part 1/3] Use TestCluster universally in sui tests (#11816)
A lot of tests spawn authority nodes individually. This is redundant to
the functionality of TestCluster.
This PR replaces a large number of those tests to use TestCluster
insdead.

There are two categories that are not update in this PR:
1. Dynamic committee change tests: these types of tests requires some
careful crafting of validator set and their keys. I need to figure out a
systematic way for integrating them into TestCluster.
2. Many of the shared object tests heavily depend on transactions.rs,
which is currently being cleaned up. Leaving that part along to avoid
merge conflicts.