Split sui-config crate (#11968)
sui-config crate started as a crate for the NodeConfig.
However it has grown too much and no longer at the bottom of the
dependency chain.
This PR splits some modules out of it:
1. Split out sui-genesis-builder. This contains all the code to build a
genesis, which depends on the execution component. After this PR,
sui-config no longer depends on the execution component, which means we
can now pass NodeConfig around even into the execution component, which
we have wanted to do for a while.
2. Split out sui-swarm-config. This contains all the configs and
builders to create a NetworkConfig for the swarm. It makes it clear now
that this is used for swarm, not for production.
3. Renamed swarm.rs into network_config.rs for clarity.
4. Renamed builder.rs into network_config_builder.rs for clarity.
5. Moved FullnodeConfigBuilder out of node.rs into
network_config_builder.rs.