[tests] Added additional programmable transactions fuzz tests (#12011)
## Description 

The idea here is to move the needle in terms of fuzz-testing
programmable transactions - exercise at least some logic throughout the
entire transaction block. Currently we measure "success" as getting to
the end of transaction block and hitting the error message about
"un-droppable" values being present at the end of the block.

## Test Plan 

This is the actual test.